+++ src/main/java/com/guerrero/Inventario/dto/AbonoCreditoDTO.java
@@ -4,6 +4,7 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Getter
@@ -13,7 +14,7 @@ import java.time.LocalDateTime;
 public class AbonoCreditoDTO {
     private Long id;
     private Long idCuenta;
-    private Double monto;
+    private BigDecimal monto;
     private LocalDateTime fecha;
     private String metodoPago;
 }
